% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit_func.R
\name{expit}
\alias{expit}
\title{Success probabilities under the logistic regression model}
\usage{
expit(beta, x, intercept = TRUE)
}
\arguments{
\item{beta}{a vector of true/estimated vector of parameters for the logistic regression model. If intercept = TRUE is specified, the first component should include the value of the intercept term}

\item{x}{a vector of covariates under the logistic regression model}

\item{intercept}{a logical argument to specify if the intercept term is included in the model or not. If TRUE, then beta should include the intercept term (default = TRUE)}
}
\value{
This function returns the calculated probability of success under the logistic regression model for the given covariate vector and vector of parameters.
}
\description{
This function calculates the success probabilities for the logistic regression model for a given vector of covariates and a vector of parameter values. The success probability for a given vector of covariates x and paramter vector beta is calculated as 1/(1 + exp(-\$beta$' X))
}
\examples{
x = c(2.5, 12.5, 7)
beta.true  = c(log(0.05), log(1.2), log(1.1), log(1.3))
# Default intercept = TRUE
prob <- expit(beta.true, x) # Calculating the probabilities

beta.true  = c(log(1.2), log(1.1), log(1.3))
# intercept = FALSE
prob <- expit(beta.true, x, intercept = FALSE) # Calculating the probabilities
}
