% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_center_lago.R
\name{mc_lago}
\alias{mc_lago}
\title{Optimum interventions under the multi center LAGO design}
\usage{
mc_lago(
  x0,
  lower,
  upper,
  nstages,
  centers,
  beta.true,
  sample.size,
  icc,
  cost.vec,
  prob,
  intercept = TRUE,
  B = 100
)
}
\arguments{
\item{x0}{a vector of starting values for the components of the intervention package}

\item{lower}{a vector providing the values of lower limits for the components of the intervention package}

\item{upper}{a vector providing the values of upper limits for the components of the intervention package}

\item{nstages}{the number of stages of the LAGO design K (>=2)}

\item{centers}{the number of centers per stage of the LAGO design}

\item{beta.true}{a vector of true i.e. population beta values}

\item{sample.size}{sample size (n) in each center of each of the K stages}

\item{icc}{the expected variation among subjects within and between the centers at any stage}

\item{cost.vec}{a vector of per unit linear cost coefficients}

\item{prob}{the desired outcome goal}

\item{intercept}{a logical argument to include intercept in the model or not (default = TRUE)}

\item{B}{the number of simulations (default = 100)}
}
\value{
This function returns a named list containing 'xopt' and 'p.opt.hat'. 'xopt' contains the stage-wise estimated optimum intervention package obtained by considering the median over the given number of simulations. 'p.opt.hat' returns the stage-wise estimated success probabilities, i.e. the success probabilities corresponding to the estimated optimal intervention package by considering the median over simulations.
}
\description{
This function takes as input several study design parameters of a multi-stage, multi-center per stage LAGO design and provides the estimated optimal intervention package through simulations. For simplicity, this function assumes an equal number of centers per stage and equal number of samples per center in each stage. The required parameters that need to be supplied to the function are an initial starting intervention package which is optimized over the stages, the lower and upper limits for the components of the intervention package, the true or best guess beta values, number of stages K, number of centers per stage J, sample size per center per stage n, the unit costs for the intervention package components and the desired outcome goal for the LAGO study. Since this function returns the estimated optimal interventions through simulations, this function also needs as input the expected variation among the subjects within and between the centers at any stage.
}
\examples{
x.init = c(2.5, 12.5, 7) # Initial value interventions
x.l = c(1, 10, 2) # Lower limits for X
x.u = c(4, 15, 15) # Upper limits for X
njk = 20 # Sample size per center per stage
K = 3 # Number of stages
J = 3 # Number of centers per stage
njk = 3 # Number of centers per stage
cost_lin = c(1, 8, 2.5) # Costs
p_bar = 0.9 # Desired outcome goal
# True/best guess beta values
beta = c(log(0.05), log(1.2), log(1.1), log(1.3))

sim_mc <- mc_lago(x0 = x.init, lower = x.l, upper = x.u,
                  beta.true = beta, nstages = K,
                  centers = J, sample.size = njk,
                  icc = 0.1, prob = p_bar,
                  cost.vec = cost_lin)

}
