% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_center_lago_unequal_centers.R
\name{mc_lago_uc}
\alias{mc_lago_uc}
\title{Optimum interventions under the multi center LAGO design with unequal centers per stage}
\usage{
mc_lago_uc(
  x0,
  lower,
  upper,
  nstages,
  centers,
  beta.true,
  sample.size,
  icc,
  bcc,
  cost.vec,
  prob,
  intercept = TRUE,
  B = 100
)
}
\arguments{
\item{x0}{a vector of starting values for the components of the intervention package}

\item{lower}{a vector providing the values of lower limits for the components of the intervention package}

\item{upper}{a vector providing the values of upper limits for the components of the intervention package}

\item{nstages}{the number of stages of the LAGO design K (>=2)}

\item{centers}{a vector of length nstages providing the number of centers at the different stages of the LAGO design}

\item{beta.true}{a vector of true i.e. population beta values}

\item{sample.size}{sample size (n) in each center of each of the K stages}

\item{icc}{the expected variation among subjects within each centers at any stage}

\item{bcc}{the expected variation between the centers at any stage while implementing the intervention package}

\item{cost.vec}{a vector of per unit linear cost coefficients}

\item{prob}{the desired outcome goal}

\item{intercept}{a logical argument to include intercept in the model or not (default = TRUE)}

\item{B}{the number of simulations (default = 100)}
}
\value{
This function returns a named list containing 'xopt' and 'p.opt.hat'. 'xopt' contains the stage-wise estimated optimum intervention package obtained by considering the median over the given number of simulations. 'p.opt.hat' returns the stage-wise estimated success probabilities, i.e. the success probabilities corresponding to the estimated optimal intervention package by considering the median over simulations.
}
\description{
Optimum interventions under the multi center LAGO design with unequal centers per stage
}
