% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lago_optimum.R
\name{opt_int}
\alias{opt_int}
\title{Optimum intervention under the LAGO design}
\usage{
opt_int(
  cost,
  beta,
  lower,
  upper,
  starting.value,
  pstar,
  intercept = TRUE,
  eps = 1e-07,
  max_eval = 3000
)
}
\arguments{
\item{cost}{a vector of linear cost coefficients}

\item{beta}{a vector of true/estimated beta values}

\item{lower}{a vector providing the values of lower limits for the components of the intervention package}

\item{upper}{a vector providing the values of upper limits for the components of the intervention package}

\item{starting.value}{a vector of starting values for the components of the intervention package}

\item{pstar}{the desired outcome goal}

\item{intercept}{a logical argument to include intercept in the model or not (default = TRUE)}

\item{eps}{the desired level of tolerance (default = 1.0e-7)}

\item{max_eval}{the maximum number of iterations to perform (default = 3000)}
}
\value{
The function returns a named list 'Optimum_Intervention' containing the value of the optimum intervention package and 'Obtained_p' containing the outcome goal under the optimal intervention package.
}
\description{
The first goal of a Learn-As-You-Go (LAGO) study is to identify the optimal intervention package while minimizing the cost of the intervention package subject to the probability of a desired binary outcome being above a given threshold. The current LAGO design considers a logistic regression model with linear cost coefficients.
This function takes as input the vector of linear cost coefficients, the true/estimated beta values, the lower and upper limits for the components of the intervention package, the desired outcome goal and a starting value of the intervention package for the optimization algorithm and calculates the optimal intervention package.
}
\examples{
# Defining vector of starting values for the algorithm
x.init = c(2.5, 12.5, 7)
# Defining vector of lower limits for the components
x.l = c(1, 10, 2)
# Defining vector of upper limits for the components
x.u = c(4, 15, 15)
# Defining vector of linear cost coefficients
cost_lin = c(1, 8, 2.5)
p_bar = 0.9 # Defining the desired outcome goal
# True beta values for the study
beta = c(log(0.05), log(1.2), log(1.1), log(1.3))
## Running the LAGO optimization algorithm
opt_lago = opt_int(cost = cost_lin, beta = beta, lower = x.l,
                   upper = x.u, pstar = p_bar, starting.value = x.init)
}
