% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_center_lago.R
\name{sc_lago}
\alias{sc_lago}
\title{Optimum package under the single center LAGO design}
\usage{
sc_lago(
  x0,
  lower,
  upper,
  nstages,
  beta.true,
  sample.size,
  icc,
  cost.vec,
  prob,
  intercept = TRUE,
  B = 100
)
}
\arguments{
\item{x0}{a vector of starting values for the components of the intervention package}

\item{lower}{a vector providing the values of lower limits for the components of the intervention package}

\item{upper}{a vector providing the values of upper limits for the components of the intervention package}

\item{nstages}{the number of stages of the LAGO design K (>=2)}

\item{beta.true}{a vector of true i.e. population beta values}

\item{sample.size}{sample size (n) in each of the K stages}

\item{icc}{the expected variation among subjects within the center at any stage}

\item{cost.vec}{a vector of per unit linear cost coefficients}

\item{prob}{the desired outcome goal}

\item{intercept}{a logical argument to include intercept in the model or not (default = TRUE)}

\item{B}{the number of simulations (default = 100)}
}
\value{

}
\description{
Optimum package under the single center LAGO design
}
